+++ lib/src/cli/options.dart
@@ -101,6 +101,7 @@ void defineOptions(ArgParser parser,
         help: 'Produce machine-readable JSON output.',
         hide: !verbose);
   }
+
   parser.addFlag('follow-links',
       negatable: false,
       help: 'Follow links to files and directories.\n'
@@ -108,6 +109,10 @@ void defineOptions(ArgParser parser,
       hide: !verbose);
   parser.addFlag('version',
       negatable: false, help: 'Show dart_style version.', hide: !verbose);
+  parser.addMultiOption('enable-experiment',
+      help: 'Enable one or more experimental features.\n'
+          'See dart.dev/go/experiments.',
+      hide: !verbose);
 
   if (verbose) parser.addSeparator('Options when formatting from stdin:');